+++ src/backtrace/unwind.rs
@@ -103,6 +103,12 @@ pub fn target(core: &mut Core, elf: &Elf, target_info: &TargetInfo) -> Output {
             // the stack was not corrupted.
             let reset_fn_range = elf.reset_fn_range();
             output.corrupted = !(reset_fn_range.contains(&pc) || reset_fn_range.is_empty());
+            if output.corrupted {
+                log::debug!(
+                    "Frame (PC {pc:#010x}) is not within reset vector ({reset_fn_range:#010x?}) \
+                            and links to itself, the stack is corrupted",
+                );
+            }
             break;
         }
 
@@ -125,6 +131,7 @@ pub fn target(core: &mut Core, elf: &Elf, target_info: &TargetInfo) -> Output {
             {
                 stacked
             } else {
+                log::debug!("Unreadable stacked exception registers, the stack is corrupted");
                 output.corrupted = true;
                 break;
             };
@@ -143,9 +150,13 @@ pub fn target(core: &mut Core, elf: &Elf, target_info: &TargetInfo) -> Output {
             break;
         }
 
-        // if the SP is above the start of the stack, the stack is corrupt
         match registers.get(registers::SP) {
             Ok(advanced_sp) if advanced_sp > target_info.stack_start => {
+                log::debug!(
+                    "SP ({advanced_sp:#010x}) above start of the stack ({:#010x}), \
+                             the stack is corrupted",
+                    target_info.stack_start
+                );
                 output.corrupted = true;
                 break;
             }