+++ src/target_info.rs
@@ -33,7 +33,6 @@ impl TargetInfo {
         elf: &Elf,
         memory_map: Vec<MemoryRegion>,
         probe_target: probe_rs::Target,
-        stack_start: u32,
     ) -> anyhow::Result<Self> {
         let active_ram_region =
             extract_active_ram_region(&probe_target, elf.vector_table.initial_stack_pointer);
@@ -46,7 +45,7 @@ impl TargetInfo {
             memory_map,
             probe_target,
             stack_info,
-            stack_start,
+            stack_start: elf.vector_table.initial_stack_pointer,
         })
     }
 }